+++ src/lib/monaco-setup.ts
@@ -1,6 +1,0 @@
-"use client";
-
-import { loader } from "@monaco-editor/react";
-import * as monaco from "monaco-editor";
-
-loader.config({ monaco });